+++ src/utilits/constant.js
@@ -1,0 +1,17 @@
+const cardsContainer = document.querySelector('.cards');
+const popupFormCreate = document.querySelector('.popup__form_belong_card');
+const buttonInfo = document.querySelector('.profile-info__edit-button');
+const popupFormInfo = document.querySelector('.popup__form_belong_profile');
+const inputName = popupFormInfo.querySelector('.popup__field_belong_name');
+const inputNote = popupFormInfo.querySelector('.popup__field_belong_note');
+const popupAvatar = document.querySelector('.popup_form_avatar');
+const buttonImage = document.querySelector('.profile__image-container');
+const buttonAdd = document.querySelector('.profile__add-button');
+
+
+
+
+
+
+
+export { cardsContainer, popupFormCreate, buttonAdd, buttonInfo, popupFormInfo, inputName, inputNote, popupAvatar, buttonImage }
